+++ src/Anonymization/Anonymizer/Pattern/Part.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MakinaCorpus\DbToolsBundle\Anonymization\Anonymizer\Pattern;
+
+interface Part {}